+++ packages/components/src/custom-select-control-v2/types.ts
@@ -51,6 +51,10 @@ export type CustomSelectButtonProps = {
 
 export type _CustomSelectProps = CustomSelectButtonProps & {
 	/**
+	 * Additional className added to the root wrapper element.
+	 */
+	className?: string;
+	/**
 	 * The child elements. This should be composed of `CustomSelectItem` components.
 	 */
 	children: React.ReactNode;