+++ packages/dataviews/src/bulk-actions-toolbar.tsx
@@ -18,28 +18,29 @@ import { useRegistry } from '@wordpress/data';
  * Internal dependencies
  */
 import { ActionWithModal } from './item-actions';
-import type { Action, AnyItem } from './types';
+import type { Action } from './types';
 import type { ActionTriggerProps } from './item-actions';
+import type { SetSelection } from './private-types';
 
-interface ActionButtonProps< Item extends AnyItem > {
+interface ActionButtonProps< Item > {
 	action: Action< Item >;
 	selectedItems: Item[];
 	actionInProgress: string | null;
 	setActionInProgress: ( actionId: string | null ) => void;
 }
 
-interface ToolbarContentProps< Item extends AnyItem > {
+interface ToolbarContentProps< Item > {
 	selection: string[];
 	actionsToShow: Action< Item >[];
 	selectedItems: Item[];
-	onSelectionChange: ( selection: Item[] ) => void;
+	onSelectionChange: SetSelection;
 }
 
-interface BulkActionsToolbarProps< Item extends AnyItem > {
+interface BulkActionsToolbarProps< Item > {
 	data: Item[];
 	selection: string[];
 	actions: Action< Item >[];
-	onSelectionChange: ( selection: Item[] ) => void;
+	onSelectionChange: SetSelection;
 	getItemId: ( item: Item ) => string;
 }
 
@@ -62,7 +63,7 @@ const SNACKBAR_VARIANTS = {
 	},
 };
 
-function ActionTrigger< Item extends AnyItem >( {
+function ActionTrigger< Item >( {
 	action,
 	onClick,
 	isBusy,
@@ -87,7 +88,7 @@ function ActionTrigger< Item extends AnyItem >( {
 
 const EMPTY_ARRAY: [] = [];
 
-function ActionButton< Item extends AnyItem >( {
+function ActionButton< Item >( {
 	action,
 	selectedItems,
 	actionInProgress,
@@ -125,13 +126,13 @@ function ActionButton< Item extends AnyItem >( {
 	);
 }
 
-function renderToolbarContent< Item extends AnyItem >(
+function renderToolbarContent< Item >(
 	selection: string[],
 	actionsToShow: Action< Item >[],
 	selectedItems: Item[],
 	actionInProgress: string | null,
 	setActionInProgress: ( actionId: string | null ) => void,
-	onSelectionChange: ( selection: Item[] ) => void
+	onSelectionChange: SetSelection
 ) {
 	return (
 		<>
@@ -179,7 +180,7 @@ function renderToolbarContent< Item extends AnyItem >(
 	);
 }
 
-function ToolbarContent< Item extends AnyItem >( {
+function ToolbarContent< Item >( {
 	selection,
 	actionsToShow,
 	selectedItems,
@@ -214,7 +215,7 @@ function ToolbarContent< Item extends AnyItem >( {
 	return buttons.current;
 }
 
-export default function BulkActionsToolbar< Item extends AnyItem >( {
+export default function BulkActionsToolbar< Item >( {
 	data,
 	selection,
 	actions = EMPTY_ARRAY,